+++ BUILD.gn
@@ -125,6 +125,20 @@ config("config") {
   }
 }
 
+# This "fuchsia_legacy" configuration includes old, non-embedder API sources and
+# defines the LEGACY_FUCHSIA_EMBEDDER symbol.  This config and its associated
+# template are both transitional and will be removed after the embedder API
+# transition is complete.
+#
+# See `source_set_maybe_fuchsia_legacy` in //flutter/common/config.gni
+#
+# TODO(fxb/54041): Remove when no longer neccesary.
+config("fuchsia_legacy") {
+  if (is_fuchsia) {
+    defines = [ "LEGACY_FUCHSIA_EMBEDDER" ]
+  }
+}
+
 config("export_dynamic_symbols") {
   if (is_linux || is_fuchsia) {
     inputs = [
@@ -151,9 +165,12 @@ if (is_fuchsia) {
 
     deps = [
       "//flutter/flow:flow_tests",
+      "//flutter/flow:flow_tests_next",
       "//flutter/fml:fml_tests",
       "//flutter/runtime:runtime_tests",
+      "//flutter/runtime:runtime_tests_next",
       "//flutter/shell/common:shell_tests",
+      "//flutter/shell/common:shell_tests_next",
       "//flutter/shell/platform/fuchsia/flutter:flutter_runner_scenic_tests",
       "//flutter/shell/platform/fuchsia/flutter:flutter_runner_tests",
     ]
