+++ flow/raster_cache.h
@@ -8,7 +8,6 @@
 #include <memory>
 #include <unordered_map>
 
-#include "flutter/flow/instrumentation.h"
 #include "flutter/flow/raster_cache_key.h"
 #include "flutter/fml/macros.h"
 #include "flutter/fml/memory/weak_ptr.h"
