+++ flow/raster_cache_unittests.cc
@@ -5,6 +5,8 @@
 #include "flutter/flow/raster_cache.h"
 
 #include "gtest/gtest.h"
+#include "third_party/skia/include/core/SkCanvas.h"
+#include "third_party/skia/include/core/SkPaint.h"
 #include "third_party/skia/include/core/SkPicture.h"
 #include "third_party/skia/include/core/SkPictureRecorder.h"
