+++ shell/common/BUILD.gn
@@ -77,7 +77,7 @@ template("dart_embedder_resources") {
   }
 }
 
-source_set("common") {
+source_set_maybe_fuchsia_legacy("common") {
   sources = [
     "animator.cc",
     "animator.h",
@@ -115,17 +115,7 @@ source_set("common") {
     "vsync_waiter_fallback.h",
   ]
 
-  deps = [
-    "//flutter/assets",
-    "//flutter/common",
-    "//flutter/flow",
-    "//flutter/fml",
-    "//flutter/lib/ui",
-    "//flutter/runtime",
-    "//flutter/shell/profiling",
-    "//third_party/dart/runtime:dart_api",
-    "//third_party/skia",
-  ]
+  public_configs = [ "//flutter:config" ]
 
   public_deps = [
     "//flutter/shell/version",
@@ -134,10 +124,30 @@ source_set("common") {
     "//third_party/rapidjson",
   ]
 
-  public_configs = [ "//flutter:config" ]
+  deps = [
+    "//flutter/assets",
+    "//flutter/common",
+    "//flutter/fml",
+    "//flutter/shell/profiling",
+    "//third_party/dart/runtime:dart_api",
+    "//third_party/skia",
+  ]
+
+  deps_legacy_and_next = [
+    "//flutter/flow:flow",
+    "//flutter/lib/ui:ui",
+    "//flutter/runtime:runtime",
+  ]
 }
 
 template("shell_host_executable") {
+  common_dep = ":common"
+  if (defined(invoker.fuchsia_legacy)) {
+    if (invoker.fuchsia_legacy) {
+      common_dep += "_fuchsia_legacy"
+    }
+  }
+
   executable(target_name) {
     testonly = true
 
@@ -148,17 +158,32 @@ template("shell_host_executable") {
     forward_variables_from(invoker, "*")
 
     deps += [
-      ":common",
-      "//flutter/fml",
       "//flutter/lib/snapshot",
-      "//flutter/runtime",
       "//flutter/runtime:libdart",
-      "//flutter/third_party/tonic",
-      "//third_party/skia",
+      common_dep,
     ]
 
     public_configs = [ "//flutter:export_dynamic_symbols" ]
   }
+}
+
+test_fixtures("shell_unittests_fixtures") {
+  dart_main = "fixtures/shell_test.dart"
+
+  fixtures = [ "fixtures/shelltest_screenshot.png" ]
+}
+
+shell_host_executable("shell_benchmarks") {
+  sources = [
+    "shell_benchmarks.cc",
+  ]
+
+  deps = [
+    ":shell_unittests_fixtures",
+    "//flutter/benchmarking",
+    "//flutter/testing:dart",
+    "//flutter/testing:testing_lib",
+  ]
 }
 
 if (enable_unittests) {
@@ -176,13 +201,7 @@ if (enable_unittests) {
     enable_metal = test_enable_metal
   }
 
-  test_fixtures("shell_unittests_fixtures") {
-    dart_main = "fixtures/shell_test.dart"
-
-    fixtures = [ "fixtures/shelltest_screenshot.png" ]
-  }
-
-  source_set("shell_test_fixture_sources") {
+  source_set_maybe_fuchsia_legacy("shell_test_fixture_sources") {
     testonly = true
 
     sources = [
@@ -197,15 +216,17 @@ if (enable_unittests) {
     ]
 
     public_deps = [
-      "//flutter/flow",
       "//flutter/fml/dart",
-      "//flutter/lib/ui",
-      "//flutter/runtime",
-      "//flutter/shell",
-      "//flutter/shell/common",
       "//flutter/testing",
-      "//flutter/testing:dart",
     ]
+
+    deps = [
+      "//flutter/assets",
+      "//flutter/common",
+      "//third_party/rapidjson",
+      "//third_party/skia",
+    ]
+
     defines = []
 
     # SwiftShader only supports x86/x64_64
@@ -216,10 +237,7 @@ if (enable_unittests) {
           "shell_test_platform_view_gl.h",
         ]
 
-        public_deps += [
-          "//flutter/shell/gpu:gpu_surface_gl",
-          "//flutter/testing:opengl",
-        ]
+        public_deps += [ "//flutter/testing:opengl" ]
 
         defines += [ "SHELL_ENABLE_GL" ]
       }
@@ -232,16 +250,29 @@ if (enable_unittests) {
       ]
 
       public_deps += [
-        "//flutter/shell/gpu:gpu_surface_vulkan",
         "//flutter/testing:vulkan",
         "//flutter/vulkan",
       ]
 
       defines += [ "SHELL_ENABLE_VULKAN" ]
     }
+
+    public_deps_legacy_and_next = [
+      "//flutter/shell/common:common",
+      "//flutter/flow:flow",
+      "//flutter/runtime:runtime",
+    ]
+
+    deps_legacy_and_next = [
+      ":shell_unittests_gpu_configuration",
+      "//flutter/lib/ui:ui",
+      "//flutter/testing:dart",
+    ]
   }
 
-  shell_host_executable("shell_unittests") {
+  source_set_maybe_fuchsia_legacy("shell_unittests_common") {
+    testonly = true
+
     sources = [
       "animator_unittests.cc",
       "canvas_spy_unittests.cc",
@@ -252,20 +283,35 @@ if (enable_unittests) {
     ]
 
     deps = [
-      ":shell_test_fixture_sources",
-      ":shell_unittests_fixtures",
-      ":shell_unittests_gpu_configuration",
       "//flutter/assets",
-      "//flutter/common",
-      "//flutter/shell",
+      "//flutter/shell/version",
     ]
 
-    if (!defined(defines)) {
-      defines = []
+    public_deps_legacy_and_next = [ ":shell_test_fixture_sources" ]
+  }
+
+  if (is_fuchsia) {
+    shell_host_executable("shell_unittests") {
+      deps = [
+        ":shell_unittests_common_fuchsia_legacy",
+        ":shell_unittests_fixtures",
+      ]
+
+      fuchsia_legacy = true
     }
 
-    if (test_enable_vulkan) {
-      defines += [ "SHELL_ENABLE_VULKAN" ]
+    shell_host_executable("shell_unittests_next") {
+      deps = [
+        ":shell_unittests_common",
+        ":shell_unittests_fixtures",
+      ]
+    }
+  } else {
+    shell_host_executable("shell_unittests") {
+      deps = [
+        ":shell_unittests_common",
+        ":shell_unittests_fixtures",
+      ]
     }
   }
 
@@ -297,18 +343,33 @@ if (enable_unittests) {
         resources += vulkan_icds
       }
     }
-  }
 
-  shell_host_executable("shell_benchmarks") {
-    sources = [
-      "shell_benchmarks.cc",
-    ]
+    fuchsia_test_archive("shell_tests_next") {
+      deps = [
+        ":shell_unittests_fixtures",
+        ":shell_unittests_next",
+      ]
 
-    deps = [
-      ":shell_unittests_fixtures",
-      "//flutter/benchmarking",
-      "//flutter/testing:dart",
-      "//flutter/testing:testing_lib",
-    ]
+      binary = "shell_unittests_next"
+
+      # TODO(gw280): https://github.com/flutter/flutter/issues/50294
+      # Right now we need to manually specify all the fixtures that are
+      # declared in the test_fixtures() call above.
+      resources = [
+        {
+          path = "$target_gen_dir/assets/kernel_blob.bin"
+          dest = "assets/kernel_blob.bin"
+        },
+        {
+          path = "$target_gen_dir/assets/shelltest_screenshot.png"
+          dest = "assets/shelltest_screenshot.png"
+        },
+      ]
+
+      if (test_enable_vulkan) {
+        libraries = vulkan_validation_libs
+        resources += vulkan_icds
+      }
+    }
   }
 }
