+++ shell/platform/fuchsia/flutter/BUILD.gn
@@ -14,10 +14,11 @@ import("//flutter/tools/fuchsia/fuchsia_libs.gni")
 import("//flutter/vulkan/config.gni")
 import("engine_flutter_runner.gni")
 
-shell_gpu_configuration("fuchsia_gpu_configuration") {
+# Fuchsia uses its own custom Surface implementation.
+shell_gpu_configuration("fuchsia_legacy_gpu_configuration") {
   enable_software = false
   enable_gl = false
-  enable_vulkan = true
+  enable_vulkan = false
   enable_metal = false
 }
 
@@ -319,10 +320,10 @@ executable("flutter_runner_unittests") {
     "//build/fuchsia/pkg:scenic_cpp",
     "//build/fuchsia/pkg:sys_cpp_testing",
     "//flutter/common",
-    "//flutter/flow",
-    "//flutter/lib/ui",
-    "//flutter/runtime",
-    "//flutter/shell/common",
+    "//flutter/flow:flow_fuchsia_legacy",
+    "//flutter/lib/ui:ui_fuchsia_legacy",
+    "//flutter/runtime:runtime_fuchsia_legacy",
+    "//flutter/shell/common:common_fuchsia_legacy",
     "//flutter/shell/platform/fuchsia/runtime/dart/utils",
     "//flutter/testing",
     "//third_party/dart/runtime:libdart_jit",
@@ -352,9 +353,10 @@ executable("flutter_runner_tzdata_unittests") {
     "//build/fuchsia/pkg:async-loop-default",
     "//build/fuchsia/pkg:scenic_cpp",
     "//build/fuchsia/pkg:sys_cpp_testing",
-    "//flutter/lib/ui",
-    "//flutter/runtime",
-    "//flutter/shell/common",
+    "//flutter/flow:flow_fuchsia_legacy",
+    "//flutter/lib/ui:ui_fuchsia_legacy",
+    "//flutter/runtime:runtime_fuchsia_legacy",
+    "//flutter/shell/common:common_fuchsia_legacy",
     "//flutter/shell/platform/fuchsia/runtime/dart/utils",
     "//flutter/testing",
     "//third_party/dart/runtime:libdart_jit",
@@ -497,16 +499,16 @@ executable("flutter_runner_scenic_unittests") {
     "//build/fuchsia/pkg:async-loop-default",
     "//build/fuchsia/pkg:scenic_cpp",
     "//build/fuchsia/pkg:sys_cpp_testing",
-    "//flutter/common:common",
-    "//flutter/flow:flow",
-    "//flutter/lib/ui",
-    "//flutter/runtime",
-    "//flutter/shell/common",
-    "//flutter/shell/platform/fuchsia/dart-pkg/fuchsia:sdk_ext",
-    "//flutter/shell/platform/fuchsia/dart-pkg/zircon:zircon",
+    "//flutter/common",
+    "//flutter/flow:flow_fuchsia_legacy",
+    "//flutter/lib/ui:ui_fuchsia_legacy",
+    "//flutter/runtime:runtime_fuchsia_legacy",
+    "//flutter/shell/common:common_fuchsia_legacy",
+    "//flutter/shell/platform/fuchsia/dart-pkg/fuchsia",
+    "//flutter/shell/platform/fuchsia/dart-pkg/zircon",
     "//flutter/shell/platform/fuchsia/runtime/dart/utils",
     "//flutter/testing",
-    "//flutter/vulkan:vulkan",
+    "//flutter/vulkan",
     "//third_party/dart/runtime:libdart_jit",
     "//third_party/dart/runtime/platform:libdart_platform_jit",
     "//third_party/icu",
